+++ mcp_examples/bin/workflow_client.dart
@@ -15,11 +15,7 @@ import 'package:google_generative_ai/google_generative_ai.dart' as gemini;
 
 /// The list of Gemini models that are accepted as a "--model" argument.
 /// Defaults to the first one in the list.
-const List<String> allowedGeminiModels = [
-  'gemini-2.5-pro-preview-03-25',
-  'gemini-2.0-flash',
-  'gemini-2.5-flash-preview-04-17',
-];
+const List<String> allowedGeminiModels = ['gemini-2.5-pro', 'gemini-2.5-flash'];
 
 void main(List<String> args) {
   final geminiApiKey = Platform.environment['GEMINI_API_KEY'];
@@ -161,8 +157,6 @@ final class WorkflowClient extends MCPClient with RootsSupport {
       StreamSinkTransformer.fromHandlers(
         handleData: (String data, EventSink<List<int>> innerSink) {
           innerSink.add(utf8.encode(data));
-          // It's a log, so we want to make sure it's always up-to-date.
-          fileByteSink.flush();
         },
         handleError: (
           Object error,
@@ -549,6 +543,13 @@ final class WorkflowClient extends MCPClient with RootsSupport {
         );
       case JsonType.bool:
         return gemini.Schema.boolean(
+          description: description,
+          nullable: nullable,
+        );
+      case JsonType.enumeration:
+        final schema = inputSchema as EnumSchema;
+        return gemini.Schema.enumString(
+          enumValues: schema.values.toList(),
           description: description,
           nullable: nullable,
         );