+++ pkgs/dart_mcp_server/test_fixtures/counter_app/lib/driver_main.dart
@@ -1,0 +1,12 @@
+// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:flutter_driver/driver_extension.dart';
+
+import 'main.dart' as app;
+
+void main() {
+  enableFlutterDriverExtension();
+  app.main();
+}